Flip default obs_dtype to fp16
OBS_CONFIG_DEFAULTS now assembles obs as fp16, so every new CUDA training run
halves the bytes on the worker->GPU handoff path (the measured n=20 pin-copy
ceiling, docs/2026-06/6.06-7). Under autocast the conv consumes fp16 directly;
the parity test confirms the fp16 build is bit-identical to the fp32 build cast
to fp16, so this is numerically free on the training path. Loaded checkpoints
keep their own recorded dtype (pre-obs_dtype ones back-fill fp32), so existing
fp32 models are unaffected.

- dataloader smoke test: assert the fp16 default and upcast obs for the
  no-autocast model forward.
- overfit_one_batch: upcast obs for its fp32 no-autocast model call.
  (bench_throughput is unaffected -- it synthesizes fp32 obs, never build_obs.)

+++ training/bc/obs_config.py
@@ -68,7 +68,7 @@ class ObsConfig:
 # by `ModelConfig.obs`'s default and by every "I want the defaults" call site.
 OBS_CONFIG_DEFAULTS = ObsConfig(
     dense_history_n=5,
-    obs_dtype="fp32",
+    obs_dtype="fp16",
 )
 
 # Default obs channel count (n=5 → 96). Convenience alias for code that builds a
